+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Examples.hs
@@ -14,11 +14,20 @@ module Test.Cardano.Ledger.Dijkstra.Examples (
 
 import Cardano.Ledger.Address (DirectDeposits (..))
 import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo)
-import Cardano.Ledger.BaseTypes (Exclusive (..), Inclusive (..), Network (..), StrictMaybe (..))
+import Cardano.Ledger.BaseTypes (
+  EpochNo (..),
+  Exclusive (..),
+  Inclusive (..),
+  Network (..),
+  SlotNo (..),
+  StrictMaybe (..),
+ )
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Rules (ConwayDELEG, ConwayDelegPredFailure (..))
+import Cardano.Ledger.Conway.TxCert (ConwayGovCert (..), Delegatee (..))
 import Cardano.Ledger.Credential (Credential (..))
+import Cardano.Ledger.DRep (DRep (..))
 import Cardano.Ledger.Dijkstra (ApplyTxError (..), DijkstraEra)
 import Cardano.Ledger.Dijkstra.Rules (DijkstraLEDGER, DijkstraMEMPOOL)
 import Cardano.Ledger.Dijkstra.Scripts (
@@ -31,6 +40,7 @@ import Cardano.Ledger.Dijkstra.TxBody (
   accountBalanceIntervalsTxBodyL,
   directDepositsTxBodyL,
   guardsTxBodyL,
+  requiredTopLevelGuardsL,
   subTransactionsTxBodyL,
  )
 import Cardano.Ledger.Dijkstra.TxCert
@@ -40,9 +50,11 @@ import Cardano.Ledger.Plutus.Data (
   dataToBinaryData,
  )
 import Cardano.Ledger.Plutus.Language (Language (..), plutusBinary)
+import Cardano.Ledger.State (StakePoolParams (sppId))
 import Control.State.Transition.Extended (Embed (..))
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
+import qualified Data.OMap.Strict as OMap
 import qualified Data.OSet.Strict as OSet
 import qualified Data.Sequence.Strict as StrictSeq
 import Lens.Micro ((%~), (&), (.~), (<>~))
@@ -52,16 +64,25 @@ import Test.Cardano.Ledger.Alonzo.Examples (
   mkAlonzoBasedLedgerExamples,
  )
 import Test.Cardano.Ledger.Babbage.Examples (exampleBabbageNewEpochState)
-import Test.Cardano.Ledger.Conway.Examples (exampleConwayBasedTxBody, mkConwayBasedExampleTx)
+import Test.Cardano.Ledger.Conway.Examples (
+  exampleAnchor,
+  exampleConwayBasedTxBody,
+  exampleProposalProcedure,
+  exampleVotingProcedures,
+  mkConwayBasedExampleTx,
+ )
 import Test.Cardano.Ledger.Core.KeyPair (mkAddr)
+import Test.Cardano.Ledger.Core.Utils (mkDummySafeHash)
 import Test.Cardano.Ledger.Dijkstra.ImpTest (exampleDijkstraGenesis)
-import Test.Cardano.Ledger.Mary.Examples (exampleMultiAssetValue)
+import Test.Cardano.Ledger.Mary.Examples (exampleMultiAsset, exampleMultiAssetValue)
 import Test.Cardano.Ledger.Plutus (alwaysSucceedsPlutus)
 import Test.Cardano.Ledger.Shelley.Examples (
   LedgerExamples (..),
   examplePayKey,
   exampleStakeKey,
   exampleStakePoolParams,
+  exampleTxIns,
+  exampleWithdrawals,
   mkKeyHash,
   mkScriptHash,
  )
@@ -118,9 +139,11 @@ mkDijkstraBasedExampleTx txBody scriptPurpose =
 mkDijkstraBasedExampleTxBody ::
   forall era.
   ( DijkstraEraTxBody era
+  , AlonzoEraTxAuxData era
   , EraTx era
   , Value era ~ MaryValue
   , EraPlutusTxInfo PlutusV4 era
+  , TxCert era ~ DijkstraTxCert era
   ) =>
   TxBody TopTx era ->
   TxBody TopTx era
@@ -134,10 +157,60 @@ mkDijkstraBasedExampleTxBody txBody =
             & datumTxOutL .~ Datum (dataToBinaryData exampleDatum)
             & referenceScriptTxOutL .~ SJust (alwaysSucceeds @'PlutusV4 3)
         ]
-    & guardsTxBodyL .~ OSet.fromList [KeyHashObj $ mkKeyHash 212, ScriptHashObj $ mkScriptHash 213]
-    & subTransactionsTxBodyL .~ mempty -- Sub-transactions require complex recursive setup
+    & guardsTxBodyL
+      .~ OSet.fromList
+        [ KeyHashObj $ mkKeyHash 211
+        , KeyHashObj $ mkKeyHash 212
+        , ScriptHashObj $ mkScriptHash 213
+        ]
+    & subTransactionsTxBodyL .~ OMap.fromFoldable [subTx]
     & directDepositsTxBodyL .~ exampleDirectDeposits
     & accountBalanceIntervalsTxBodyL .~ exampleAccountBalanceIntervals
+  where
+    subTx =
+      mkBasicTx @era subTxBody
+        & witsTxL
+          <>~ ( mkBasicTxWits
+                  & scriptTxWitsL
+                    .~ Map.singleton
+                      (hashScript @era $ alwaysSucceeds @'PlutusV4 3)
+                      (alwaysSucceeds @'PlutusV4 3)
+              )
+        & auxDataTxL
+          %~ fmap
+            ( \auxData ->
+                auxData
+                  & plutusScriptsTxAuxDataL
+                    <>~ Map.singleton PlutusV4 (NE.singleton (plutusBinary (alwaysSucceedsPlutus @'PlutusV4 3)))
+            )
+    subTxBody =
+      mkBasicTxBody
+        & inputsTxBodyL .~ exampleTxIns
+        & referenceInputsTxBodyL .~ exampleTxIns
+        & outputsTxBodyL
+          <>~ StrictSeq.fromList
+            [ mkBasicTxOut
+                (mkAddr examplePayKey exampleStakeKey)
+                (exampleMultiAssetValue 2)
+                & datumTxOutL .~ Datum (dataToBinaryData exampleDatum)
+                & referenceScriptTxOutL .~ SJust (alwaysSucceeds @'PlutusV4 3)
+            ]
+        & vldtTxBodyL .~ ValidityInterval (SJust (SlotNo 2)) (SJust (SlotNo 4))
+        & mintTxBodyL .~ exampleMultiAsset 1
+        & withdrawalsTxBodyL .~ exampleWithdrawals
+        & auxDataHashTxBodyL .~ SJust (TxAuxDataHash $ mkDummySafeHash @EraIndependentTxAuxData 30)
+        & scriptIntegrityHashTxBodyL .~ SJust (mkDummySafeHash 42)
+        & networkIdTxBodyL .~ SJust Mainnet
+        & certsTxBodyL .~ exampleDijkstraCerts
+        & votingProceduresTxBodyL .~ exampleVotingProcedures
+        & proposalProceduresTxBodyL .~ OSet.singleton exampleProposalProcedure
+        & currentTreasuryValueTxBodyL .~ SJust (Coin 867530900000)
+        & treasuryDonationTxBodyL .~ Coin 1000000
+        & requiredTopLevelGuardsL
+          .~ Map.fromList
+            [ (KeyHashObj $ mkKeyHash 212, SNothing)
+            , (ScriptHashObj $ mkScriptHash 213, SJust $ exampleDatum @era)
+            ]
 
 exampleDirectDeposits :: DirectDeposits
 exampleDirectDeposits =
@@ -160,7 +233,28 @@ exampleAccountBalanceIntervals =
 
 exampleDijkstraCerts :: StrictSeq.StrictSeq (DijkstraTxCert era)
 exampleDijkstraCerts =
-  -- TODO should I add the new certs here?
   StrictSeq.fromList
-    [ DijkstraTxCertPool (RegPool exampleStakePoolParams)
+    [ DijkstraTxCertDeleg $ DijkstraRegCert (KeyHashObj (mkKeyHash 1)) (Coin 2000000)
+    , DijkstraTxCertDeleg $ DijkstraUnRegCert (KeyHashObj (mkKeyHash 1)) (Coin 2000000)
+    , DijkstraTxCertDeleg $
+        DijkstraDelegCert (KeyHashObj (mkKeyHash 3)) (DelegStake (sppId exampleStakePoolParams))
+    , DijkstraTxCertDeleg $
+        DijkstraDelegCert (KeyHashObj (mkKeyHash 4)) (DelegVote (DRepKeyHash (mkKeyHash 5)))
+    , DijkstraTxCertDeleg $
+        DijkstraDelegCert
+          (KeyHashObj (mkKeyHash 6))
+          (DelegStakeVote (sppId exampleStakePoolParams) (DRepKeyHash (mkKeyHash 7)))
+    , DijkstraTxCertDeleg $
+        DijkstraRegDelegCert
+          (KeyHashObj (mkKeyHash 8))
+          (DelegStake (sppId exampleStakePoolParams))
+          (Coin 2000000)
+    , DijkstraTxCertPool $ RegPool exampleStakePoolParams
+    , DijkstraTxCertPool $ RetirePool (sppId exampleStakePoolParams) (EpochNo 2)
+    , DijkstraTxCertGov $ ConwayRegDRep (KeyHashObj (mkKeyHash 9)) (Coin 500000) (SJust exampleAnchor)
+    , DijkstraTxCertGov $ ConwayUnRegDRep (KeyHashObj (mkKeyHash 9)) (Coin 500000)
+    , DijkstraTxCertGov $ ConwayUpdateDRep (KeyHashObj (mkKeyHash 9)) (SJust exampleAnchor)
+    , DijkstraTxCertGov $
+        ConwayAuthCommitteeHotKey (KeyHashObj (mkKeyHash 10)) (KeyHashObj (mkKeyHash 11))
+    , DijkstraTxCertGov $ ConwayResignCommitteeColdKey (KeyHashObj (mkKeyHash 10)) (SJust exampleAnchor)
     ]